refactor: split API routes into dedicated module for improved organization
